+++ src/crates/infrastructure/src/circle_duplicate_checker.rs
@@ -6,18 +6,18 @@ use domain::{
 use sqlx::MySqlPool;
 
 #[derive(Clone, Debug)]
-pub struct CircleDuplicateCheckerWithMySql {
+pub struct CircleDuplicateChecker {
     db: MySqlPool,
 }
 
-impl CircleDuplicateCheckerWithMySql {
+impl CircleDuplicateChecker {
     pub fn new(db: MySqlPool) -> Self {
         Self { db }
     }
 }
 
 #[async_trait::async_trait]
-impl CircleDuplicateCheckerInterface for CircleDuplicateCheckerWithMySql {
+impl CircleDuplicateCheckerInterface for CircleDuplicateChecker {
     async fn check_circle_duplicate(&self, circle: &Circle) -> Result<(), Error> {
         let query = "SELECT * FROM circles WHERE name = ?";
         let record = sqlx::query(query)
