+++ src/crates/infrastructure/src/circle_reader.rs
