+++ src/crates/infrastructure/src/lib.rs
@@ -1,4 +1,4 @@
-pub mod circle_duplicate_checker_with_my_sql;
-pub mod circle_reader_with_my_sql;
-pub mod circle_repository_with_my_sql;
+pub mod circle_duplicate_checker;
+pub mod circle_reader;
+pub mod circle_repository;
 pub(crate) mod maria_db_schema;
